Added check to availability filtering to ensure information has not expired
