+++ package.json
@@ -21,9 +21,10 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "aws-lambda": "^1.0.6",
     "aws-sdk": "^2.747.0",
     "axios": "^0.20.0",
+    "date-fns": "^2.17.0",
+    "date-fns-tz": "^1.1.2",
     "dotenv": "^8.2.0"
   },
   "devDependencies": {
